Do not shutdown when using architectures that use custom "make -j"

+++ _install.sh
@@ -64,4 +64,4 @@ mkdir root
 	mail SUCCEDED s3.amazonaws.com/$user/linux/$file
 	exit 0
 ) || mail FAILED
-shutdown -h now
+[[ ${arch/-j} == $arch ]] || shutdown -h now
